Refactor Repository de Libros

+++ Backend/app/Http/Controllers/Libro/EliminarLibroController.php
@@ -4,41 +4,41 @@ namespace App\Http\Controllers\Libro;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use App\Models\Libro;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Facades\Storage;
+use App\Repositories\Libro\LibroRepositoryInterface;
+
 
 class EliminarLibroController extends Controller
 {
-    /**
-     * Handle the incoming request.
-     */
-    public function __invoke(Request $request, $id)
-    {
-        //findOrFail return exception
-        try {
-            //Eliminar el libro
-            $libro=Libro::findOrFail($id);
-            $rutaPortada = $libro->portada_path;
-            $libro->delete();
+    public function __construct(
+        private readonly LibroRepositoryInterface $librosRepository
+    ){}
 
-            //Eliminar la ruta de la portada asociada, si la tiene
-            if ($rutaPortada) {
-                Storage::disk('local')->delete($rutaPortada);
-            }
-            return response()->json(
-                [   "data"=>null,
-                    "message"=>"Libro eliminado",
-                    "errors"=>[],
-                ], 204
-            );
-        //Manejo de excepcion
-        } catch (ModelNotFoundException) {
-             return response()->json([
+    public function __invoke(Request $request, $id){
+
+        $libro = $this->librosRepository->getById($id);
+
+        if(is_null($libro)){
+            return response()->json([
                 'data'=>null,
                 'message'=>'Libro no encontrado',
                 'errors'=>[]
             ], 404 );
         }
+
+        $rutaPortada = $libro->portada_path;
+
+        $libro->delete();
+
+        //Eliminar la ruta de la portada asociada, si la tiene
+        if ($rutaPortada) {
+            Storage::disk('local')->delete($rutaPortada);
+        }
+        return response()->json([
+            "data"=>null,
+            "message"=>"Libro eliminado",
+            "errors"=>[],
+        ], 204);
+
     }
 }
